requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-dotenv==1.4.2
django-embed-video==1.4.10
django-extensions==3.2.3
djangorestframework
filelock==3.14.0
idna==3.7
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
whitenoise==6.6.0
